refactor to use ts
