--- src/app/redux/providers/actions.ts
@@ -1,13 +1,0 @@
-import { createAction } from '@reduxjs/toolkit';
-import { IMessage, SEND_MESSAGE, FETCH_PROVIDERS_REQUEST, IChatActionTypes } from './types';
-
-export const fetchProvidersRequest = createAction(FETCH_PROVIDERS_REQUEST);
-
-export function fetchProvidersRequest(): ICommonProvider[] {
-  return {
-    type: DELETE_MESSAGE,
-    meta: {
-      timestamp,
-    },
-  };
-}
