--- src/app/redux/providers/index.ts
@@ -1,6 +1,0 @@
-import commonReducer from './reducers';
-export default commonReducer;
-
-// TODO: Discuss how would we like our directory structure to look for Redux? The redux docs have a 3 recommended approaches:
-
-// https://redux.js.org/faq/code-structure#what-should-my-file-structure-look-like-how-should-i-group-my-action-creators-and-reducers-in-my-project-where-should-my-selectors-go
