+++ src/app/store.ts
@@ -1,8 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { rootReducer } from './reducers';
+
+import rootReducer from './rootReducer';
 
 const store = configureStore({
   reducer: rootReducer,
 });
+
+export type AppDispatch = typeof store.dispatch;
 
 export default store;
